Move CarPlay DateComponentFormatter extensions to shared extension file in main Nav SDK. Use these extensions in Route Annotation Anotations and CPTrip. Fix new source files that were omitted from the Example-CarPlay application target.
